requirements.txt
python-dotenv
colored
psycopg2-binary
pandas
markdown
bs4
tqdm
nltk
